hids_client/requirements.txt
python-crontab==2.5.1
kafka-python
pyyaml